Proof-of-concept seems to work... enough!
Project started this morning in the trains from Rambouillet to
Saint-Cloud. Should I thank SNCF for such a mess in Versailles
Chantiers that gave me 2 hours of hacking?

Anyway, the prompt is still quite primitive, but seems to be worth
continuing to improve this daemonized prompt engine.

This is a daemon, but not a nasty demon. So it is an angel. Like a
guardian angel, attached to the shell and who gives him precious
information about the world around that you can't see. Always around,
but always discreet.