Travis-CI: do more tests on perl 5.8
Our builder (the 'dist' script) is made for perl 5.10, so we can not run
it under perl 5.8 on Travis-CI, so many tests fails.
So we won't use the builder and its result unless we are on perl 5.10+
and we will test the other things using "perl -Ilib bin/angel-PS1".
The built script is tested by Travis-CI on the 'release' branch anyway.