Escape wildcard '*' in case block of the bash code (GH #17)
As the shell code we generate is evaluated in $( ... ), it is subject to
word splitting and pathname expansion. In the code of the angel function
we have a case...esac block which has the *) entry. This 'word' was
expanded if the current directory contained a filename ending with ')'.
For example:
    touch '(aaa GH17)'
    eval $(angel-PS1)

So we use a trick to escape pathname expansion. The line:
    *)
becomes:
    ''*)